@@ -1,6 +1,6 @@
 {
   "name": "nx-electron",
-  "version": "10.0.0-beta.2",
+  "version": "11.0.0",
   "description": "Electron Plugin for Nx",
   "main": "index.js",
   "types": "index.d.ts",
@@ -9,10 +9,10 @@
   "private": true,
   "scripts": {
     "lint": "jshint .",
-    "test": "jest --maxWorkers=1 .*.spec.js",
+    "test": "jest --maxWorkers=1 .*.spec.ts",
     "compile": "tsc",
     "build": "npm run clear:all && tsc && npm run copy:all && npm run permission",
-    "publish": "cd build && npm publish --tag beta",
+    "publish": "cd build && npm publish",
     "permission": "node ./scripts/make_build_public.js",
     "copy:all": "npm run copy:ts && npm run copy:json && npm run copy:readme && npm run copy:licence && npm run copy:ignore",
     "copy:ts": "shx cp -rf src build/",
@@ -49,66 +49,54 @@
     "requirements": {},
     "migrations": "./migrations.json"
   },
-  "jest": {
-    "modulePathIgnorePatterns": [
-      "tmp",
-      "<rootDir>/test",
-      "<rootDir>/src",
-      "collection/.*/files"
-    ],
-    "testPathIgnorePatterns": [
-      "node_modules"
-    ],
-    "coverageReporters": [
-      "html"
-    ],
-    "coverageDirectory": "coverage"
-  },
   "peerDependencies": {
     "@nrwl/workspace": "*"
   },
   "dependencies": {
-    "@angular-devkit/architect": "^0.1001.6",
-    "@angular-devkit/build-webpack": "^0.1001.6",
-    "@angular-devkit/core": "^10.1.6",
-    "@angular-devkit/schematics": "^10.1.6",
-    "@nrwl/jest": "^10.3.0",
-    "@nrwl/linter": "^10.3.0",
-    "@nrwl/workspace": "^10.3.0",
-    "circular-dependency-plugin": "^5.0.2",
-    "copy-webpack-plugin": "^5.1.2",
+    "@angular-devkit/architect": "~0.1100.1",
+    "@angular-devkit/build-webpack": "~0.1100.1",
+    "@angular-devkit/core": "~11.0.1",
+    "@angular-devkit/schematics": "~11.0.1",
+    "@nrwl/jest": "^11.2.6",
+    "@nrwl/linter": "^11.2.6",
+    "circular-dependency-plugin": "^5.2.2",
+    "copy-webpack-plugin": "^6.4.1",
     "electron-builder": "^22.9.1",
-    "fork-ts-checker-webpack-plugin": "^5.2.0",
-    "license-webpack-plugin": "^2.3.0",
+    "fork-ts-checker-webpack-plugin": "^6.1.0",
+    "license-webpack-plugin": "^2.3.11",
     "rimraf": "^3.0.2",
     "source-map-support": "^0.5.19",
     "strip-json-comments": "^3.1.1",
-    "terser-webpack-plugin": "^3.1.0",
+    "terser-webpack-plugin": "^4.2.3",
     "tree-kill": "^1.2.2",
-    "ts-loader": "^7.0.5",
-    "tsconfig-paths-webpack-plugin": "3.2.0",
-    "webpack": "^4.44.2",
-    "webpack-dev-server": "^3.11.0",
-    "webpack-merge": "^4.2.2",
-    "webpack-node-externals": "1.7.2"
+    "ts-loader": "^8.0.14",
+    "tsconfig-paths-webpack-plugin": "^3.3.0",
+    "webpack": "^4.46.0",
+    "webpack-dev-server": "^3.11.2",
+    "webpack-merge": "^5.7.3",
+    "webpack-node-externals": "^2.5.2"
   },
   "devDependencies": {
+    "@nrwl/devkit": "^11.2.6",
+    "@nrwl/workspace": "^11.2.6",
     "@types/circular-dependency-plugin": "^5.0.1",
-    "@types/copy-webpack-plugin": "^5.0.2",
-    "@types/jest": "^26.0.14",
-    "@types/node": "^14.11.8",
+    "@types/copy-webpack-plugin": "^6.4.0",
+    "@types/jest": "^26.0.20",
+    "@types/node": "^14.14.22",
     "@types/rimraf": "^3.0.0",
-    "@types/terser-webpack-plugin": "^3.0.0",
-    "@types/webpack": "^4.41.22",
+    "@types/terser-webpack-plugin": "^4.2.0",
+    "@types/webpack": "^4.41.26",
     "@types/webpack-merge": "^4.1.5",
-    "@types/webpack-node-externals": "^1.7.1",
-    "bufferutil": "^4.0.1",
+    "@types/webpack-node-externals": "^2.5.0",
+    "bufferutil": "^4.0.3",
     "canvas": "^2.6.1",
-    "electron": "^9.3.2",
-    "jest": "^26.5.3",
-    "prettier": "^2.1.2",
+    "electron": "^9.4.2",
+    "jest": "^26.6.3",
+    "jshint": "^2.12.0",
+    "prettier": "^2.2.1",
     "shx": "^0.3.3",
-    "typescript": "^3.9.7",
-    "utf-8-validate": "^5.0.2"
+    "ts-jest": "^26.5.2",
+    "typescript": "^4.1.3",
+    "utf-8-validate": "^5.0.4"
   }
 }
