@@ -10,9 +10,10 @@ import { getSourceRoot } from '../../utils/workspace';
 import { normalizePackgingOptions } from '../../utils/normalize';
 
 import { Observable, from, of } from 'rxjs';
-import { map, concatMap, catchError } from 'rxjs/operators';
+import { map, tap, concatMap, catchError } from 'rxjs/operators';
 import { platform } from 'os';
 
+import { stripIndents } from '@angular-devkit/core/src/utils/literals';
 import stripJsonComments from 'strip-json-comments';
 
 try {
@@ -42,7 +43,14 @@ export default createBuilder<JsonObject & PackageElectronBuilderOptions>(run);
 
 function run(rawOptions: JsonObject & PackageElectronBuilderOptions, context: BuilderContext): Observable<PackageElectronBuilderOutput> { 
   return from(getSourceRoot(context)).pipe(
-    map(sourceRoot =>
+    tap(_ => {
+      context.logger.warn(stripIndents`
+        *********************************************************
+        DO NOT FORGET TO REBUILD YOUR FORNTEND & BACKEND PROJECTS
+        FOR PRODUCTION BEFORE PACKAGING / MAKING YOUR ARTIFACT!
+        *********************************************************`);
+    }),
+    map(({ sourceRoot, projectRoot }) =>
       normalizePackgingOptions(rawOptions, context.workspaceRoot, sourceRoot)
     ),
     map(options => 
@@ -129,7 +137,7 @@ function _createBaseConfig(options: PackageElectronBuilderOptions, context: Buil
       {
           from: resolve(options.sourcePath, options.name),
           to: options.name,
-          filter: ['main.js', 'assets']
+          filter: ['main.js', 'preload.js', 'assets']
       },
       {
           from: resolve(options.sourcePath, options.name),
