@@ -1,6 +1,6 @@
-export const nxVersion = '^10.3.0';
-export const nxElectronVersion = '^9.0.0';
-export const electronVersion = '^11.0.3';
+export const nxVersion = '^11.2.6';
+export const nxElectronVersion = '^11.0.0';
+export const electronVersion = '^11.3.0';
 export const electronBuilderVersion = '^22.9.1';
 export const rimrafVersion = '^3.0.0';
 export const exitZeroVersion = '^1.0.1';
