@@ -1,20 +1,19 @@
 import { BuilderContext } from '@angular-devkit/architect';
-import { workspaces } from '@angular-devkit/core';
-import { NodeJsSyncHost } from '@angular-devkit/core/node';
+import { workspaces, normalize } from '@angular-devkit/core';
+import { NxScopedHost } from '@nrwl/devkit/ngcli-adapter';
 
-export async function getSourceRoot(context: BuilderContext) {
-    const workspaceHost = workspaces.createWorkspaceHost(new NodeJsSyncHost());
-    const { workspace } = await workspaces.readWorkspace(
-      context.workspaceRoot,
-      workspaceHost
-    );
+export async function getSourceRoot(context: BuilderContext): Promise<{ sourceRoot: string; projectRoot: string }> {
+  const workspaceHost = workspaces.createWorkspaceHost(new NxScopedHost(normalize(context.workspaceRoot)));
+  const { workspace } = await workspaces.readWorkspace('', workspaceHost);
+  const { project } = context.target;
+  const { sourceRoot, root } = workspace.projects.get(project);
+
+  if (sourceRoot && root) {
+    return { sourceRoot, projectRoot: root };
+  }
   
-    if (workspace.projects.get(context.target.project).sourceRoot) {
-      return workspace.projects.get(context.target.project).sourceRoot;
-    } else {
-      context.reportStatus('Error');
-      const message = `${context.target.project} does not have a sourceRoot. Please define one.`;
-      context.logger.error(message);
-      throw new Error(message);
-    }
+  context.reportStatus('Error');
+  const message = `${project} does not have a sourceRoot or root. Please define both.`;
+  context.logger.error(message);
+  throw new Error(message);
 }
